bug fix for google API OAuth2 process / change procfile for single process environment
