+++ server/Procfile
@@ -1,2 +1,1 @@
 web: node ./server.js
-worker: node ./worker.js
