+++ server/controllers/jobPostingFetcher.js
@@ -7,7 +7,7 @@ const jobposting = db.jobposting
 const {logger} = require('../config/logger')
 require('dotenv').config()
 
-const MaxPageToProbe = 15 // In case of not using page limit, please put undefined or comment this line out
+const MaxPageToProbe = 10 // In case of not using page limit, please put undefined or comment this line out
 
 
 function sliceIntoChunks(arr, chunkSize) {
