+++ server/util/taskScheduler/jobpostingFetchScheduler.js
@@ -12,7 +12,7 @@ require('dotenv').config();
 // and will return error in case of repeated incoming queries at short interval
 
 const { NODE_ENV } = process.env;
-
+logger.info(`[Bull jpProcessQueue] NODE_ENV ${NODE_ENV} `)
 const cronOpt = (() => {
     if(NODE_ENV === 'test' || NODE_ENV === 'development')
     {
@@ -20,7 +20,7 @@ const cronOpt = (() => {
     }
     else{
         // At 10:00 UTC on Monday, Tuesday, Wednesday, Thursday, and Friday.
-        return { cron : '0 17 * * 1,2,3,4,5'};
+        return { cron : '11 17 * * 1,2,3,4,5'};
     }
 })();
